requirements.txt
fastapi
uvicorn
python-socketio
pymongo
python-dotenv